Final Commit (Hopefully)
Fixed datetime bug with MySQL Schema. Last modified times are now recorded as time since last epoch across s3 and native filesystem roots. Changed   behavior of functions to suit new multi-root system. Reminder: a configuration file is required to log into a mysql server so the program can perform actions. Database is assumed to be static, but can still be changed with the --db argument. As long as both the username and password of the mysql server account are the same as the other server, the program should be able to log in. You can also specify a different configuration file with --config. The --root argument is now deprecated since there are multiple roots from now on. The argument will now just list the initial root stored in metadata. Currently there is no safety for adding and deleting roots (--addroot, --delroot). The --dups and --jreport function still only give results for the last run scan (regardless of root).